==> scripts/3_bold_mergeall.Rmd <==
Read bold data and unified taxonomic information

```{r message=FALSE, echo=FALSE, warning=FALSE}
bold_data<-read.csv("../data/bold/bold_data.csv")
infotax <- read.csv("../data/taxonomic_info_all.csv")
```

```{r message=FALSE, echo=FALSE, warning=FALSE}
#group taxonomic information on the last columns of the dataframe
#move columns 9-24 to the end of the data frame
bold<-cbind(bold_data[,1:8],bold_data [,25:81], bold_data[,9:24])

#delete columns with extra taxonomic ids
bold<-bold[,-(c(66,68,70,72,74,76,78,80))]
#assign kingdom to rows with all na except kingdom
bold[which(!is.na(bold[,66]) & is.na(bold[,67]) & is.na(bold[,68]) & is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),65]<-infotax[match(bold[which(!is.na(bold[,66]) & is.na(bold[,67]) & is.na(bold[,68]) & is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),66], infotax$phylum_name), 1 ]
#assign kingdom and phylum with all na except class
bold[which(!is.na(bold[,67]) & is.na(bold[,68]) & is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),65:66]<-infotax[match(bold[which(!is.na(bold[,67]) & is.na(bold[,68]) & is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),67], infotax$class_name), 1:2]
#assign kingdom, phylum and class to rows with all na except order
bold[which(!is.na(bold[,68]) & is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),65:67]<-infotax[match(bold[which(!is.na(bold[,68]) & is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),68], infotax$order_name), 1:3]

# assign kingdom, phylum, class and order to rows with all NA except family
bold[which(!is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),65:68]<-infotax[match(bold[which(!is.na(bold[,69]) & is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),69], infotax$family_name), 1:4]

#assign kingdom, phylum class, order and family to rows with all NA except subfamily
bold[which(!is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),65:69]<-infotax[match(bold[which(!is.na(bold[,70]) & is.na(bold[,71]) & is.na(bold[,72])),70], infotax$subfamily_name), 1:5]
# assign kingdom, phylum, class, order, family and subgamily to rows with all NA except genus
bold[which(!is.na(bold[,71]) & is.na(bold[,72])),65:70]<-infotax[match(bold[which(!is.na(bold[,71]) & is.na(bold[,72])),71], infotax$genus_name), 1:6]

#assign all taxonomic information matching by species
bold[which(!is.na(bold[,72])),65:71]<-infotax[match(bold[which(!is.na(bold[,72])),72], infotax$species_name), 1:7]

#add column tax_id
bold$Tax_id<-NA
bold[which(!is.na(bold[,72])),74]<-infotax[match(bold[which(!is.na(bold[,72])),72], infotax$species_name), 9]

```


```{r message=FALSE, echo=FALSE, warning=FALSE}
write_csv(bold,"../data/bold/bold_complete.csv")
```
